+++ src/app/(mainapp)/blog/ayam-1/page.mdx
